fix: create product
