+++ internal/delivery/http/v1/controller/product/product_controller.go
@@ -3,8 +3,10 @@ package productController
 import (
 	productHandler "eniqlo/internal/delivery/http/v1/controller/product/handler"
 	productService "eniqlo/internal/service/product"
+	cryptoJWT "eniqlo/package/crypto/jwt"
 
 	"github.com/go-playground/validator/v10"
+	echojwt "github.com/labstack/echo-jwt/v4"
 	"github.com/labstack/echo/v4"
 )
 
@@ -12,8 +14,8 @@ func Init(group *echo.Group, val *validator.Validate, productSvc productService.
 	user := group.Group("/product")
 	handler := productHandler.NewHandler(productSvc, val)
 
-	publicRoute := user
-	publicRoute.Use()
-	publicRoute.POST("", handler.CreateProduct)
-	publicRoute.PUT("/:id", handler.CreateProduct)
+	privateRoute := user
+	privateRoute.Use(echojwt.WithConfig(cryptoJWT.JWTConfig()))
+	privateRoute.POST("", handler.CreateProduct)
+	privateRoute.PUT("/:id", handler.CreateProduct)
 }
